Redistribute the read tests in the translation units
